+++ lib/capway_sync/reactor/v1/steps/compare_data_v2.ex
@@ -1,5 +1,6 @@
 defmodule CapwaySync.Reactor.V1.Steps.CompareDataV2 do
   alias ElixirSense.Log
+  alias CapwaySync.Models.Dynamodb.ActionItem
   use Reactor.Step
   require Logger
 
@@ -23,36 +24,38 @@ defmodule CapwaySync.Reactor.V1.Steps.CompareDataV2 do
     Logger.info("Starting data comparison between Trinity and Capway subscribers")
 
     Logger.info(
-      "Trinity active subscriber mapset: #{inspect(trinity_subscriber_data.active_subscribers.map_set)}"
+      "Trinity active subscriber mapset: #{inspect(trinity_subscriber_data.active_subscribers)}"
     )
 
     Logger.info(
-      "Capway active subscriber mapset: #{inspect(capway_subscriber_data.active_subscribers.map_set)}"
+      "Capway active subscriber mapset: #{inspect(capway_subscriber_data.active_subscribers)}"
     )
 
     cancel_contracts =
       get_contracts_to_cancel(
-        capway_subscriber_data.active_subscribers.data,
-        trinity_subscriber_data.active_subscribers.map_set
+        capway_subscriber_data.active_subscribers,
+        trinity_subscriber_data.active_subscribers
       )
 
     update_contracts =
       get_contracts_to_update(
-        capway_subscriber_data.active_subscribers.data,
-        trinity_subscriber_data.active_subscribers.data
+        capway_subscriber_data.active_subscribers,
+        trinity_subscriber_data.active_subscribers
       )
 
     create_contracts =
       get_contracts_to_create(
-        trinity_subscriber_data.active_subscribers.data,
-        capway_subscriber_data.active_subscribers.map_set
+        trinity_subscriber_data.active_subscribers,
+        capway_subscriber_data.active_subscribers
       )
 
-    suspend_accounts =
-      get_accounts_to_suspend(
-        capway_subscriber_data.inactive_subscribers.data,
-        trinity_subscriber_data.active_subscribers.map_set
+    {suspend_accounts, cancel_accounts_2} =
+      get_accounts_to_suspend_or_cancel(
+        capway_subscriber_data.inactive_subscribers,
+        trinity_subscriber_data.active_subscribers
       )
+
+    cancel_accounts = Map.merge(cancel_accounts_2, cancel_contracts)
 
     data = %{
       trinity: %{
@@ -72,25 +75,43 @@ defmodule CapwaySync.Reactor.V1.Steps.CompareDataV2 do
   end
 
   @doc """
-   This function identifies Capway contracts that needs to be suspended in Trinity.
+   This function identifies Capway contracts that needs to be either suspended or cancelled in Trinity.
+   The current rule is based on whetever the trinity subscription is locked in or not. Ie if it has a time based contract.
+   If it is locked in, it should be suspended, otherwise cancelled.
   """
 
-  def get_accounts_to_suspend(capway_subscriber_data, trinity_subscriber_map_set) do
-    Enum.reduce(capway_subscriber_data, [], fn capway_sub, acc ->
-      if MapSet.member?(trinity_subscriber_map_set, capway_sub.trinity_subscriber_id) do
-        # Subscriber should be suspended in trinity
-        [
-          %{
-            national_id: capway_sub.national_id,
-            trinity_subscriber_id: capway_sub.trinity_subscriber_id,
-            reason: "Should be suspended in Trinity"
-          }
-          | acc
-        ]
-      else
-        acc
+  def get_accounts_to_suspend_or_cancel(capway_subscriber_data, trinity_subscriber_data) do
+    suspend = %{}
+    cancellations = %{}
+
+    Enum.each(capway_subscriber_data, fn {_id, capway_sub} ->
+      if Map.has_key?(trinity_subscriber_data, capway_sub.trinity_subscriber_id) do
+        {_id, trinity_sub} = Map.get(trinity_subscriber_data, capway_sub.trinity_subscriber_id)
+
+        if trinity_sub.subscription_type == "locked" do
+          # Should be suspended
+          item =
+            ActionItem.create_action_item(:suspend, %{
+              national_id: capway_sub.national_id,
+              trinity_subscriber_id: capway_sub.trinity_subscriber_id,
+              reason: "Should be suspended in Trinity due to locked subscription"
+            })
+
+          Map.put(suspend, capway_sub.trinity_subscriber_id, item)
+        else
+          item =
+            ActionItem.create_action_item(:trinity_cancel_subscription, %{
+              national_id: capway_sub.national_id,
+              trinity_subscriber_id: capway_sub.trinity_subscriber_id,
+              reason: "Should be cancelled in Trinity due to inactive Capway status"
+            })
+
+          Map.put(cancellations, capway_sub.trinity_subscriber_id, item)
+        end
       end
     end)
+
+    {suspend, cancellations}
   end
 
   @doc """
@@ -98,24 +119,24 @@ defmodule CapwaySync.Reactor.V1.Steps.CompareDataV2 do
    It will focus on checking for all active Trinity subscribers
    and see if they exist in Capway data, if not they will be marked for creation.
   """
-  def get_contracts_to_create(trinity_subscriber_data, capway_subscriber_map_set) do
-    Enum.reduce(trinity_subscriber_data, [], fn trinity_sub, acc ->
-      if MapSet.member?(
-           capway_subscriber_map_set,
-           trinity_sub.trinity_subscriber_id
+  def get_contracts_to_create(trinity_subscriber_data, capway_subscriber_data) do
+    Enum.reduce(trinity_subscriber_data, %{}, fn {trinity_subscriber_id, trinity_sub}, acc ->
+      if Map.has_key?(
+           capway_subscriber_data,
+           trinity_subscriber_id
          ) do
         # Subscriber exists in both systems, no action needed
         acc
       else
         # Subscriber missing in Capway, mark for creation
-        [
-          %{
+        item =
+          ActionItem.create_action_item(:capway_create_contract, %{
             national_id: trinity_sub.national_id,
-            trinity_subscriber_id: trinity_sub.trinity_subscriber_id,
+            trinity_subscriber_id: trinity_subscriber_id,
             reason: "Missing in Capway system"
-          }
-          | acc
-        ]
+          })
+
+        Map.put(acc, trinity_sub.trinity_subscriber_id, item)
       end
     end)
   end
@@ -128,56 +149,50 @@ defmodule CapwaySync.Reactor.V1.Steps.CompareDataV2 do
   """
   def get_contracts_to_update(capway_subscriber_data, trinity_subscriber_data) do
     # First filter out the Capway subscribers that have no national_id
-    Enum.reduce(capway_subscriber_data, [], fn capway_sub, acc ->
-      # Start by checking if the subscriber exists in Trinity data
-      # trinity_sub =
-      #   Enum.find(trinity_subscriber_data, fn trinity_sub ->
-      #     trinity_sub.trinity_subscriber_id == capway_sub.trinity_subscriber_id
-      #   end)
-
-      with trinity_sub <-
-             Enum.find(trinity_subscriber_data, fn ts ->
-               ts.trinity_subscriber_id == capway_sub.trinity_subscriber_id
-             end),
+    Enum.reduce(capway_subscriber_data, %{}, fn {_id, capway_sub}, acc ->
+      with true <- Map.has_key?(trinity_subscriber_data, capway_sub.trinity_subscriber_id),
+           {_id, trinity_sub} <-
+             Map.get(
+               trinity_subscriber_data,
+               capway_sub.trinity_subscriber_id
+             ),
            false <- trinity_sub == nil,
            false <- trinity_sub.national_id == capway_sub.national_id do
         # National IDs are different, mark for update
-        [
-          %{
+        item =
+          ActionItem.create_action_item(:capway_update_contract, %{
             national_id: trinity_sub.national_id,
             trinity_subscriber_id: capway_sub.trinity_subscriber_id,
             reason: "National ID mismatch"
-          }
-          | acc
-        ]
+          })
+
+        Map.put(acc, capway_sub.trinity_subscriber_id, item)
       else
         _ -> acc
       end
     end)
   end
 
-  def get_contracts_to_cancel(capway_subscriber_data, trinity_subscriber_map_set) do
-    Enum.reduce(capway_subscriber_data, [], fn capway_sub, acc ->
-      if MapSet.member?(
-           trinity_subscriber_map_set,
-           capway_sub.trinity_subscriber_id
+  def get_contracts_to_cancel(capway_subscriber_data, trinity_subscriber_data) do
+    Enum.reduce(capway_subscriber_data, %{}, fn {trinity_subscriber_id, capway_sub}, acc ->
+      if Map.has_key?(
+           trinity_subscriber_data,
+           trinity_subscriber_id
          ) do
         # Subscriber exists in both systems, no action needed
         acc
       else
-        Logger.info(
-          "Marking Capway subscriber #{capway_sub.trinity_subscriber_id} for cancellation"
-        )
+        Logger.info("Marking Capway subscriber #{trinity_subscriber_id} for cancellation")
 
         # Subscriber missing in Trinity, mark for cancellation
-        [
-          %{
+        item =
+          ActionItem.create_action_item(:capway_cancel_contract, %{
             national_id: capway_sub.national_id,
-            trinity_subscriber_id: capway_sub.trinity_subscriber_id,
+            trinity_subscriber_id: trinity_subscriber_id,
             reason: "Missing in Trinity system"
-          }
-          | acc
-        ]
+          })
+
+        Map.put(acc, trinity_subscriber_id, item)
       end
     end)
   end